+++ threedigrid_builder/grid/fwrapper/quadtree.pyx
@@ -51,6 +51,7 @@ cpdef create_quadtree(
     int[:] mmax,
     int[:] nmax,
     int lgrmin,
+    int use_2d_flow,
     short[::1,:] area_mask,
     int[::1,:] quad_idx,
     int[::1,:] lg,
@@ -69,6 +70,7 @@ cpdef create_quadtree(
         mmax=&mmax[0],
         nmax=&nmax[0],
         lgrmin=&lgrmin,
+        use_2d_flow=&use_2d_flow,
         area_mask=&area_mask[0,0],
         lg=&lg[0,0],
         quad_idx=&quad_idx[0,0],
@@ -78,7 +80,7 @@ cpdef create_quadtree(
         i1=&i1,
         n_cells=&n_cells,
         n_line_u=&n_line_u,
-        n_line_v=&n_line_v
+        n_line_v=&n_line_v,
     )
 
     return <object>n_cells, <object>(n_line_u, n_line_v)